+++ R/generics.R
@@ -1,3 +1,4 @@
+#' @export
 print.kdecopula <- function(x, ...) {
     cat("Kernel copula density estimate")
     ## add variable names if available
@@ -8,6 +9,10 @@ print.kdecopula <- function(x, ...) {
     invisible(x)
 }
 
+
+
+#' @importFrom stats logLik
+#' @export
 summary.kdecopula <- function(object, ...) {
     cat("Kernel copula density estimate\n")
     cat("------------------------------\n")
@@ -117,11 +122,12 @@ summary.kdecopula <- function(object, ...) {
 #' AIC(dens.est)
 #' BIC(dens.est)
 #'
+#' @export
 logLik.kdecopula <- function(object, ...) {
     if (!is.null(object$info)) {
         ## access info slot if available
         out <- object$info$loglik
-        effp   <- object$info$effp
+        effp <- object$info$effp
     } else {
         ## calculate log likelihood and effp from data
         likvalues <- dkdecop(object$udata, object)
@@ -131,9 +137,6 @@ logLik.kdecopula <- function(object, ...) {
                             object$bw,
                             object$method,
                             object$lfit)
-        # for TLL effp is stored already
-        if (object$method %in% c("TLL1", "TLL2"))
-            effp <- object$effp
     }
 
     ## add attributes
@@ -147,11 +150,13 @@ logLik.kdecopula <- function(object, ...) {
 
 expand_method <- function(method) {
     switch(method,
-           "TLL1" = "Transformation local likelihood, log-linear ('TLL1')",
-           "TLL2" = "Transformation local likelihood, log-quadratic ('TLL2')",
-           "T"    = "Transformation estimator ('T')",
-           "MR"   = "Mirror-reflection ('MR')",
-           "beta" = "Beta kernels ('beta')",
-           "TTPI" = "Tapered transformation estimator (plug-in)",
-           "TTCV" = "Tapered transformation estimator (cross-validated)")
+           "TLL1"   = "Transformation local likelihood, log-linear ('TLL1')",
+           "TLL2"   = "Transformation local likelihood, log-quadratic ('TLL2')",
+           "TLL1nn" = "Transformation local likelihood, log-linear (nearest-neighbor, 'TLL1nn')",
+           "TLL2nn" = "Transformation local likelihood, log-quadratic (nearest-neighbor, 'TLL2nn')",
+           "T"      = "Transformation estimator ('T')",
+           "MR"     = "Mirror-reflection ('MR')",
+           "beta"   = "Beta kernels ('beta')",
+           "TTPI"   = "Tapered transformation estimator (plug-in, 'TTPI')",
+           "TTCV"   = "Tapered transformation estimator (cross-validated, 'TTCV')")
 }