+++ R/kdecop.R
@@ -8,21 +8,18 @@
 #'   
 #' @param udata \code{nx2} matrix of copula data.
 #' @param bw bandwidth specification; if \code{NA}, \code{bw} is selected
-#' automatically; Otherwise, please provide (for the respective method) \cr
-#' \code{"MR", "beta"}: a positive real number, \cr
-#' \code{"T"}: a \eqn{2x2} matrix for method, \cr
-#' \code{"TLL1", "TLL2"}: a list with (named) entries \code{B} and \code{alpha}
-#' containing the \eqn{2x2} rotation matrix \code{B} and the nearest-neighbor 
-#' fraction \code{alpha}, \cr
+#' automatically (see Details); Otherwise, please provide \cr
+#' \code{"T", "TLL1", "TLL2"}: a \eqn{2x2} bandwidth matrix, \cr
+#' \code{"TLL1nn", "TLL2nn"}: a list with (named) entries \code{B}, \code{alpha},
+#' and \code{kappa}, \cr
 #' \code{"TTCV", "TTPI"}: a numeric vector of length four containing \eqn{(h, 
-#' \rho, \theta_1, \theta_2)}, c.f. Wen and Wu (2015).
+#' \rho, \theta_1, \theta_2)}, c.f. Wen and Wu (2015), \cr
+#' \code{"MR", "beta"}: a positive real number.
 #' @param mult bandwidth multiplier, has to be positive; useful for making 
 #' estimates more/less smooth manually.
 #' @param method 
-#' \code{"MR"}: mirror-reflection estimator, \cr 
-#' \code{"beta"}: beta kernel estimator, \cr 
-#' \code{"T"}: transformation estimator with standard bivariate kernel 
-#' estimation, \cr 
+#' \code{"T"}: transformation estimator based on classical bivariate kernel 
+#' estimation (e.g., Geenenens et al., 2014), \cr 
 #' \code{"TLL1"}: transformation estimator with log-linear local likelihood
 #' estimation (Geenenens et al., 2014), \cr 
 #' \code{"TLL2"}: transformation estimator with log-quadradtic local likelihood
@@ -31,9 +28,14 @@
 #' estimation and nearest-neighbor bandwidths (Geenenens et al., 2014), \cr 
 #' \code{"TLL2nn"}: transformation estimator with log-quadradtic local likelihood 
 #' estimation and nearest-neighbor bandwidths (Geenenens et al., 2014), \cr
-#' \code{"TTPI"}: tapered transformation estimator with plug-in bandwidths, \cr
+#' \code{"TTPI"}: tapered transformation estimator with plug-in bandwidths
+#' (Wu and Wen, 2015), \cr
 #' \code{"TTCV"}: tapered transformation estimator with profile cross-validation
-#' bandwidths.
+#' bandwidths (Wu and Wen, 2015), \cr
+#' \code{"MR"}: mirror-reflection estimator (Gijbels and Mielniczuk, 1990), \cr 
+#' \code{"beta"}: beta kernel estimator (Charpentier et al., 2006), \cr
+#' \code{"bern"}: Bernstein copula estimator (Sanchetta and Satchell, 2004); the
+#' coefficients are adjusted by the method of Weiss and Scheffer (2012).
 #' @param knots integer; number of knots in each dimension for the spline
 #' approximation.
 #' @param renorm.iter integer; number of iterations for the renormalization
@@ -61,19 +63,21 @@
 #' options are available with \code{\link[kdecopula:plot.kdecopula]{plot}}
 #' and \code{\link[kdecopula:contour.kdecopula]{contour}}.
 #' 
-#' @details Details on the estimation methods and bandwidth selection can be
-#' found in Geenens et al. (2014) for methods \code{TLL1/2nn} and Nagler (2014) 
-#' for other methods. For \code{TLL1/2}, we use the rule of thumb
-#' \deqn{cov(\Phi^{-1}(udata))^{1/2} * 3.5 * n^{1 / (4 * deg + 2)}}.  
-#' We use a Gaussian product kernel function for all methods 
-#' except the beta kernel estimator.\cr 
+#' @details We use a Gaussian product kernel function for all methods 
+#' except the beta kernel and Bernstein estimators. For details on bandwidth 
+#' selection for a specific method, see: \code{\link{bw_t}}, 
+#' \code{\link{bw_tll}}, \code{\link{bw_tll_nn}}, \code{\link{bw_tt_pi}}, 
+#' \code{\link{bw_tt_cv}}, \code{\link{bw_mr}}, \code{\link{bw_beta}}, 
+#' \code{\link{bw_bern}}.
+#' \cr 
 #' 
 #' Kernel estimates are usually no proper copula densities. In particular, the
-#' estimated marginal densities are not uniform. We mitigate this issue bei
-#' implementing a renormalization procedure. The number of iterations of the
+#' estimated marginal densities are not uniform. We mitigate this issue by
+#' a renormalization procedure. The number of iterations of the
 #' renormalization algorithm can be specified with the \code{renorm.iter}
 #' argument. Typically, a very small number of iterations is sufficient. \cr
 #' 
+#' @note 
 #' The implementation of the tapered transformation estimator ("TTPI"/"TTCV") 
 #' was kindly provided by Kuangyu Wen. 
 #' 
@@ -91,16 +95,28 @@
 #' Probit transformation for nonparametric kernel estimation of the copula
 #' density.
 #' arXiv:1404.4414 [stat.ME]. 
-#' \cr \cr 
-#' Nagler, T. (2014). 
-#' Kernel Methods for Vine Copula Estimation.
-#' Master's Thesis, Technische Universitaet Muenchen,
-#' \url{https://mediatum.ub.tum.de/node?id=1231221} 
 #' \cr \cr
 #' Wen, K. and Wu, X. (2015).
 #' Transformation-Kernel Estimation of the Copula Density,
 #' Working paper,
 #' \url{http://agecon2.tamu.edu/people/faculty/wu-ximing/agecon2/public/copula.pdf}
+#' \cr \cr
+#' Gijbels, I. and Mielniczuk, J. (1990).
+#' Estimating the density of a copula function.
+#' Communications in Statistics - Theory and Methods, 19(2):445-464. 
+#' \cr \cr 
+#' Charpentier, A., Fermanian, J.-D., and Scaillet, O. (2006).
+#' The estimation of copulas: Theory and practice. 
+#' In Rank, J., editor, Copulas: From theory to application in finance. Risk Books.
+#' \cr \cr 
+#' Weiss, G. and Scheffer, M. (2012).
+#' Smooth Nonparametric Bernstein Vine Copulas.
+#' arXiv:1210.2043 [q-fin.RM]
+#' \cr \cr
+#' Nagler, T. (2014). 
+#' Kernel Methods for Vine Copula Estimation.
+#' Master's Thesis, Technische Universitaet Muenchen,
+#' \url{https://mediatum.ub.tum.de/node?id=1231221} 
 #' 
 #' @examples
 #' 
@@ -109,150 +125,170 @@
 #' udat <- apply(wdbc[, -1], 2, function(x) rank(x)/(length(x)+1))
 #' 
 #' ## estimation of copula density of variables 5 and 6
-#' dens.est <- kdecop(udat[, 5:6])
-#' summary(dens.est)
-#' plot(dens.est) 
+#' est <- kdecop(udat[, 5:6])
+#' summary(est)
+#' plot(est) 
+#' contour(est)
 #' 
 #' ## evaluate density estimate at (u1,u2)=(0.123,0.321)
-#' dkdecop(c(0.123, 0.321), dens.est) 
+#' dkdecop(c(0.123, 0.321), est) 
 #' 
 #' ## evaluate cdf estimate at (u1,u2)=(0.123,0.321)
-#' pkdecop(c(0.123, 0.321), dens.est) 
+#' pkdecop(c(0.123, 0.321), est) 
 #' 
 #' ## simulate 500 samples from density estimate
-#' plot(rkdecop(500, dens.est))  # pseudo-random
-#' plot(rkdecop(500, dens.est), quasi = TRUE)  # quasi-random
+#' plot(rkdecop(500, est))  # pseudo-random
+#' plot(rkdecop(500, est), quasi = TRUE)  # quasi-random
 #' 
-kdecop <- function(udata, bw = NA, mult = 1, method = "TLL2", knots = 30, renorm.iter = 3L, info = TRUE) {
+#' 
+#' @import lattice 
+#' @importFrom grDevices col2rgb colorRampPalette rgb
+#' @importFrom graphics contour plot
+#' 
+#' @export
+#' 
+kdecop <- function(udata, bw = NA, mult = 1, method = "TLL2", knots = 30, 
+                   renorm.iter = 3L, info = TRUE) {
     udata <- as.matrix(udata)
-    n <- nrow(udata)
-    d <- ncol(udata)
+    n <- NROW(udata)
+    d <- NCOL(udata)
+    check_kdecop_input(as.list(environment()))
     
-    ## sanity checks
-    if (n < 2)
-        stop("Number of observations has to be at least 2.")
-    if (d != 2)
-        stop("Dimension has to be 2.")
-    if (any(udata > 1) | any(udata < 0))
-        stop("'udata' have to be in the interval [0,1].")
-    if (!(method %in% c("MR", "beta", "T",
-                        "TLL1", "TLL2", "TLL1nn", "TLL2nn",
-                        "TTPI", "TTCV", "bern")))
-        stop("method not implemented")
-    if (mult <= 0)
-        stop("'mult' has to be a positive number.")
-    if (is.na(knots))
-        knots <- 100/d
-    knots <- round(knots)
-    stopifnot(is.numeric(renorm.iter))
-    renorm.iter <- round(renorm.iter)
-    stopifnot(renorm.iter >= 0)
-    stopifnot(is.logical(info))
-    
-    ## bandwidth selection and adjustment (with bandwidth checks)
-    if (missing(bw))
-        bw <- bw_select(udata, method)
+    # bandwidth selection
     if (any(is.na(bw)))
         bw <- bw_select(udata, method)
-    if (method %in% c("TLL1nn", "TLL2nn")) {
-        if (is.null(bw$B) | is.null(bw$alpha))
-            stop("For methods 'TLL1/2nn', you have to provide a list 'bw = list(B = <your.B>,  alpha = <your.alpha>)'
-where both parts of the bandwidth specification are provided via  'your.B', and 'your.alpha'.")
-        if (any(c(diag(bw$B), bw$alpha) <= 0))
-            stop("Bandwidths have to be positive.")
-        bw$alpha <- mult * bw$alpha
-    } else if (method %in% c("TTPI", "TTCV")) {
-        if (bw[1] < 0)
-            stop("The smoothing parameter (bw[1]) has to be positive.")
-        if (bw[3] < 0)
-            stop("The first tapering parameter (bw[3]) has to be positive.")
-        if (abs(bw[2] > 0.9999))
-            stop("The correlation parameter (bw[2]) has to lie in (-1,1).")
-        bw[1] <- bw[1] * mult
-    } else if (method == "T") {
-        B <- as.matrix(bw)
-        if (nrow(B) == 1)
-            bw <- diag(as.numeric(bw), d)
-        bw <- mult * bw
-        bw <- bw * sqrt(min((n^(-1/6))^2 / det(bw), 1))
-    } else if (method == "MR") {
-        bw <- min(bw * mult, 1)
-    } else if (method == "beta") {
-        bw <- bw * mult
-    } else if (method == "bern") {
-        bw <- round(bw * mult)
-    }
+    bw <- multiply_bw(bw, mult, method, d)
+    check_bw(bw, method)
     
-    ## construct grid with k knots in dimension d
-    knots <- 30
-    pnts <- pnorm(seq(-3.25, 3.25, l = knots))
-    grid <- as.matrix(do.call(expand.grid,
-                              split(rep(pnts, d), rep(c(1, 2), each = knots))))
-    
-    ## fit model for method TLL
-    if (method %in% c("TLL1nn", "TLL2nn")) {
-        zdata <-  qnorm(udata)
-        lfit <- my_locfit(zdata,
-                          bw$B,
-                          bw$alpha,
-                          bw$kappa,
-                          deg = as.numeric(substr(method, 4, 4)),
-                          grid = grid)
-    } else if (method %in% c("TLL1", "TLL2")) {
-        zdata <-  qnorm(udata)
-        lfit <- my_locfitc(zdata,
-                           bw,
-                           mult,
-                           deg = as.numeric(substr(method, 4, 4)),
-                           grid = grid)
+    # fit model for method TLL
+    if (method %in% c("TLL1", "TLL2", "TLL1nn", "TLL2nn")) {
+        lfit <- my_locfit(udata, bw, deg = as.numeric(substr(method, 4, 4)))
     } else {
         lfit <- NULL
     }
     
-    ## evaluate estimator on grid
-    evalf <- eval_func(method)  # get evaluation function
+    # evaluate estimator on grid
+    grid <- make_grid(knots, d)
+    evalf <- eval_func(method)
     object <- list(udata = udata,
                    bw = bw,
                    lfit = lfit,
                    method = method)
-    vals <- array(evalf(grid, obj = object), dim = rep(knots, d))
+    vals <- array(evalf(grid$expanded, obj = object), dim = rep(knots, d))
+    rm("object")
+    # rescale copula density to have uniform margins
+    vals <- renorm2unif(vals, grid, renorm.iter)
     
-    ## rescale copula density to have uniform margins
-    if (renorm.iter > 0) {
-        tmplst <- split(rep(seq.int(knots)-1, d-1),
-                        ceiling(seq.int(knots*(d-1))/knots))
-        helpind <- as.matrix(do.call(expand.grid, tmplst))
-        vals <- renorm(vals, pnts, renorm.iter, helpind)
-    }
-    
-    ## store results
+    # store results
     res <- list(udata    = udata,
-                grid     = pnts,
+                grid     = grid$pnts,
                 estimate = vals,
                 bw       = bw,
                 mult     = mult,
                 method   = method)
     class(res) <- "kdecopula"
+    # add further (information if asked for)
+    with_fit_info(res, info, lfit)
+}
+
+#' Sanity checks for kdecop function
+#'
+#' @param args list of all argument-value pairs
+#'
+#' @noRd
+check_kdecop_input <- function(args) {
+    if (args$n < 2)
+        stop("Number of observations has to be at least 2.")
+    if (args$d != 2)
+        stop("Dimension has to be 2.")
+    if (any(args$udata > 1) | any(args$udata < 0))
+        stop("'udata' have to be in the interval [0,1].")
     
-    ## add effp for TLL
-    if (method %in% c("TLL1", "TLL2", "TLL1nn", "TLL2nn"))
-        res$effp <- eff_num_par(udata, likvalues, bw, method, lfit)
+    if (!(args$method %in% c("MR", "beta", "T",
+                             "TLL1", "TLL2", "TLL1nn", "TLL2nn",
+                             "TTPI", "TTCV", "bern")))
+        stop("method not implemented")
     
-    ## add further information if asked for
+    if (args$mult <= 0)
+        stop("'mult' has to be a positive number.")
+    
+    stopifnot(is.numeric(args$knots))
+    if (!all.equal(args$knots, as.integer(args$knots)))
+        stop("'knots' must be a positive integer.")
+    if (args$knots < 1)
+        stop("'knots' must be a positive integer.")
+    
+    stopifnot(is.numeric(args$renorm.iter))
+    if (!all.equal(args$renorm.iter, as.integer(args$renorm.iter)))
+        stop("'renorm.iter' must be a positive integer.")
+    if (args$renorm.iter < 0)
+        stop("'renorm.iter' must be a non-negative integer.")
+    
+    stopifnot(is.logical(args$info))
+}
+
+#' Grid for spline representation of the copula density estimate
+#'
+#' @param knots number of knots
+#' @param d dimension
+#' @noRd
+make_grid <- function(knots, d) {
+    pnts <- pnorm(seq(-3.25, 3.25, l = knots))
+    pntlst <- split(rep(pnts, d), rep(c(1, 2), each = knots))
+    expanded <- as.matrix(do.call(expand.grid, pntlst))
+    list(pnts = pnts, expanded = expanded)
+}
+
+#' Normalizing the estimate to uniform margins
+#'
+#' @param vals initial values of the copula density evaluated on the grid
+#' @param grid a list with expanded grid and univarite gridpoints
+#' @param renorm.iter integer; number of iterations for the renormalization
+#' procedure
+#'
+#' @return normalized values on the grid
+#'
+#' @noRd
+renorm2unif <- function(vals, grid, renorm.iter) {
+    if (renorm.iter > 0) {
+        d <- NCOL(grid$expanded)
+        knots <- length(grid$pnts)
+        tmplst <- split(rep(seq.int(knots)-1, d-1),
+                        ceiling(seq.int(knots*(d-1))/knots))
+        helpind <- as.matrix(do.call(expand.grid, tmplst))
+        vals <- renorm(vals, grid$pnts, renorm.iter, helpind)
+    }
+    vals
+}
+
+#' Finalize kdecopula object
+#'
+#' @param res kdecopula object
+#' @param info logical; whether further information about the fit shall be 
+#' computed
+#' @param lfit locfit object (only for TLL methods, otherwise NULL)
+#'
+#' @return the kdecopula object with added info entry (if demanded).
+#'
+#' @noRd
+with_fit_info <- function(res, info, lfit) {
     if (info) {
         # log-likelihood
-        likvalues <- dkdecop(udata, res)
+        likvalues <- dkdecop(res$udata, res)
         loglik <- sum(log(likvalues))
-        if (!(method %in% c("TTPI", "TTCV"))) {
+        if (!(res$method %in% c("TTPI", "TTCV", "bern"))) {
             # effective number of parameters
-            effp <- eff_num_par(udata, likvalues, bw, method, lfit)
+            effp <- eff_num_par(res$udata,
+                                likvalues, 
+                                res$bw,
+                                res$method, 
+                                lfit)
             # information criteria
             AIC  <- - 2 * loglik + 2 * effp
+            n <- nrow(res$udata)
             cAIC <- AIC + (2 * effp * (effp + 1)) / (n - effp - 1)
             BIC  <- - 2 * loglik + log(n) * effp
         } else {
-            #             warning("Effective number of parameters not yet implemented for this method.
-            # Use 'info = FALSE' if you don't want to see this message.")
             effp <- AIC <- cAIC <- BIC <- NA
         }
         
@@ -265,6 +301,5 @@ where both parts of the bandwidth specification are provided via  'your.B', and 
                          BIC       = BIC)
     }
     
-    ## return results as kdecopula object
     res
 }